# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface ResetPasswordRequest {
  newPassword: string;
}

export interface SecurityGroupSummary {
  id: number;
  name: string;
}

export interface User {
  id: number;
  firstName: string;
  surname: string;
  email: string;
  isAdmin: boolean;
  isActive: boolean;
  securityGroups: SecurityGroupSummary[];
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  /** JWT token for mobile clients */
  token: string;
}

export interface CreateUserRequest {
  firstName: string;
  surname: string;
  email: string;
  password: string;
  isAdmin?: boolean;
  isActive?: boolean;
  securityGroupIds?: number[];
}

export interface UpdateUserRequest {
  firstName?: string;
  surname?: string;
  email?: string;
  isAdmin?: boolean;
  isActive?: boolean;
  securityGroupIds?: number[];
}

export interface UserSummary {
  id: number;
  firstName: string;
  surname: string;
  email: string;
}

export interface SecurityGroup {
  id: number;
  name: string;
  members: UserSummary[];
  createdAt: string;
}

export interface CreateSecurityGroupRequest {
  name: string;
  memberIds?: number[];
}

export interface UpdateSecurityGroupRequest {
  name?: string;
  memberIds?: number[];
}

export interface Category {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreateCategoryRequest {
  name: string;
}

export interface Priority {
  id: number;
  name: string;
  order: number;
  createdAt: string;
}

export interface CreatePriorityRequest {
  name: string;
  order: number;
}

export interface Task {
  id: number;
  name: string;
  description?: string | null;
  parentId?: number | null;
  categoryId?: number | null;
  categoryName?: string | null;
  priorityId?: number | null;
  priorityName?: string | null;
  ownerId?: number | null;
  ownerName?: string | null;
  assignedToId?: number | null;
  assignedToName?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  dueDate?: string | null;
  startDate?: string | null;
  hasChildren: boolean;
  childCount: number;
  noteCount: number;
  securityGroupIds: number[];
  securityGroups: SecurityGroupSummary[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskRequest {
  name: string;
  description?: string | null;
  parentId?: number | null;
  categoryId?: number | null;
  priorityId?: number | null;
  ownerId?: number | null;
  assignedToId?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  dueDate?: string | null;
  startDate?: string | null;
  securityGroupIds?: number[];
}

export interface UpdateTaskRequest {
  name?: string;
  description?: string | null;
  categoryId?: number | null;
  priorityId?: number | null;
  ownerId?: number | null;
  assignedToId?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  dueDate?: string | null;
  startDate?: string | null;
  securityGroupIds?: number[];
}

export interface TaskNote {
  id: number;
  taskId: number;
  content: string;
  authorId: number;
  authorName: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskNoteRequest {
  content: string;
}

export interface DashboardSummary {
  totalTasks: number;
  completedTasks: number;
  overdueTasks: number;
  myOpenTasks: number;
  tasksInProgress: number;
  upcomingDueSoon: number;
}

export interface TaskCountByGroup {
  id?: number | null;
  name: string;
  count: number;
}

export type ListUsersParams = {
search?: string;
active?: boolean;
};

export type ListTasksParams = {
/**
 * Filter by parent task (null for root tasks)
 */
parentId?: number | null;
ownerId?: number;
assignedToId?: number;
categoryId?: number;
priorityId?: number;
search?: string;
/**
 * Return only root-level tasks (no parent)
 */
rootOnly?: boolean;
};

export type GetMyTasksParams = {
limit?: number;
};