# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /security-groups:
    get:
      operationId: listSecurityGroups
      tags: [security-groups]
      summary: List all security groups
      responses:
        "200":
          description: List of security groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityGroup"
    post:
      operationId: createSecurityGroup
      tags: [security-groups]
      summary: Create a security group (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecurityGroupRequest"
      responses:
        "201":
          description: Security group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityGroup"
  /security-groups/{id}:
    get:
      operationId: getSecurityGroup
      tags: [security-groups]
      summary: Get a security group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Security group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityGroup"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSecurityGroup
      tags: [security-groups]
      summary: Update a security group (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecurityGroupRequest"
      responses:
        "200":
          description: Updated security group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityGroup"
    delete:
      operationId: deleteSecurityGroup
      tags: [security-groups]
      summary: Delete a security group (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Categories
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List task categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a task category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Priorities
  /priorities:
    get:
      operationId: listPriorities
      tags: [priorities]
      summary: List task priorities
      responses:
        "200":
          description: List of priorities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Priority"
    post:
      operationId: createPriority
      tags: [priorities]
      summary: Create a task priority (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriorityRequest"
      responses:
        "201":
          description: Priority created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Priority"
  /priorities/{id}:
    put:
      operationId: updatePriority
      tags: [priorities]
      summary: Update a priority (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriorityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Priority"
    delete:
      operationId: deletePriority
      tags: [priorities]
      summary: Delete a priority (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Tasks
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT token for mobile clients
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        securityGroups:
          type: array
          items:
            $ref: "#/components/schemas/SecurityGroupSummary"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - surname
        - email
        - isAdmin
        - isActive
        - securityGroups
        - createdAt
...
    SecurityGroupSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    SecurityGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - members
        - createdAt
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
      required:
        - id
        - firstName
        - surname
        - email
    CreateSecurityGroupRequest:
      type: object
      properties:
        name:
          type: string
        memberIds:
          type: array
          items:
            type: integer
      required:
        - name
    UpdateSecurityGroupRequest:
      type: object
      properties:
        name:
          type: string
        memberIds:
          type: array
          items:
            type: integer
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Priority:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - order
        - createdAt
    CreatePriorityRequest:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
      required:
        - name
        - order
...