# lib/api-zod/src/generated/types/priority.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Priority {
  id: number;
  name: string;
  order: number;
  createdAt: Date;
}