# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SecurityGroupSummary } from './securityGroupSummary';

export interface User {
  id: number;
  firstName: string;
  surname: string;
  email: string;
  isAdmin: boolean;
  isActive: boolean;
  securityGroups: SecurityGroupSummary[];
  createdAt: Date;
}