# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Task Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  id: number;
  firstName: string;
  surname: string;
  email: string;
}